fix(rs): cells type mismatch

+++ game-of-life/with-wasm/src/lib.rs
@@ -104,7 +104,7 @@ impl Universe {
         self.height
     }
 
-    pub fn cells(&self) -> *const u32 {
+    pub fn cells(&self) -> *const usize {
         self.cells.as_slice().as_ptr()
     }
 }
